+++ lib/map/cached_tile_provider.dart
@@ -5,9 +5,9 @@ import 'package:flutter_map/flutter_map.dart';
 class CachedTileProvider extends TileProvider {
   CachedTileProvider();
   @override
-  ImageProvider getImage(Coords<num> coords, TileLayer options) {
+  ImageProvider getImage(TileCoordinates coordinates, TileLayer options) {
     return CachedNetworkImageProvider(
-      getTileUrl(coords, options),
+      getTileUrl(coordinates, options),
       //Now you can set options that determine how the image gets cached via whichever plugin you use.
     );
   }